backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
qdrant-client
sentence-transformers
pymupdf
python-multipart
groq